Initial completion of Rock Paper Scissors
Made game fucntion, for playing 5 rounds of the game
Made a confirm function for the user if they want to play the game
Made an alert for the scoreboard for the user to know what the score is.
Polished the pChoice function for setting the user choice into the variable
Made a retry function to retry the game and reset the scoreboard.
Polished getComputerChoice function to alert what the computer choice is.

!!! Note: This was for the previous commit because I commited it blank.