Add UI to the rock-paper-scissor project
Learned how to create another git branch.
Created the new branch 'rps-ui' so that new changes won't affect the
already working project.
Downloaded and added the new images for the choices.
Added Java DOM manipulation elements for the scoreboard.
Added Java eventlistener to the button and use the game functions.
edited the style.css to add style to the UI