Finnish rps-ui project as per TOP Java funda
Revised the index.html for restructuring, added the header, footer
and for DOM targets.
Added Java DOM manipulation for context and instruction.
Revised all the prompts() and confirm() into a context Java DOM.
Added variable 'i' for rounds counter and be able to reset the game.
Added Java DOM manipulation for button that restart the game.
Revised eventlistener of player choice for a good flow of the game.
Revised the game function for proper calling of other functions.
Added win() function for counting the rounds and announce the winner.
Added retry() function for restarting the game.
Revised and added some style for highlighting and presentation.